+++ pika/include/pika_define.h
@@ -7,7 +7,7 @@
 #define PIKA_DEFINE_H_
 
 
-#define PIKA_MAX_WORKER_THREAD_NUM 24
+#define PIKA_MAX_WORKER_THREAD_NUM 128
 
 const std::string kPikaPidFile = "pika.pid";