+++ pika/third/slash/slash/include/env.h
@@ -70,6 +70,7 @@ bool GetDescendant(const std::string& dir, std::vector<std::string>& result);
 
 
 uint64_t NowMicros();
+int64_t GetCurrentTime();
 void SleepForMicroseconds(int micros);
 
 Status NewSequentialFile(const std::string& fname, SequentialFile** result);
@@ -83,6 +84,8 @@ Status AppendSequentialFile(const std::string& fname, SequentialFile** result);
 Status AppendWritableFile(const std::string& fname, WritableFile** result, uint64_t write_len = 0);
 
 Status NewRandomRWFile(const std::string& fname, RandomRWFile** result);
+
+int SetSysMinFreeKbytesRatio(const double ratio);
 
 // Clear system cached memory
 int ClearSystemCachedMemory();